test: update name of tests
